Backend/Inference/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.20
pillow==11.3.0
numpy>=2.0,<2.4
requests==2.32.4
pytest==8.4.1
httpx==0.28.1
onnxruntime==1.24.2
python-dotenv==1.1.1
pydantic==2.5.0
